Refactor code from main2 to utils
